event-manager/pyproject.toml
[project]
name = "event-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.47",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]
